bring config comments more in line with docs
I'm writing a tool which may be taking this in to generate
documentation, and I'm using srchd as a base since it's one of my very
few completed projects that I get a lot of use of.
I'm lazy and don't want to write good high quality documentation only
for it to get out of sync with the application itself; why not generate
the documentation from the source code?

There is a non-zero chance none of this means anything and the only
reference to it will be here.
I'm still trying to prototype something.