introduce a far better method of testing engines
The previous method relied solely on someone checking to see if the
engine's output was sane or not.
Of course, nobody did, and undeterministic testing is a truly horrible
idea, especially when the network is involved.

This new test harness, engtest, relies on someone updating test cases
periodically instead, but makes testing infinitely easier when you have
known inputs and outputs.
I will be updating all engines to use engtest and adding the testdata
that results from it to Git soon; if it is a good idea to do so or not,
I have no idea, but I will anyway.

The engtest code is not particularly great, but it does the job.