ddg, google, yahoo: new tests
The search queries were chosen arbitrarily.

For transparency, this commit used to contain search/engines/testdata
but I have amended this commit to not do so; including huge blobs of
data for testing in srchd's repository doesn't sound like a great idea
at this time.
I think it would be best if the test cases remain private for the time
being until I figure out something better.
Contact me if you are interested in the tests.