ui: improve styling of index search box
On smaller screens it would look ugly, and the distance between the
footer to the search bar was different from the distance to the header
to the search bar.
While it didn't bother me because I never look at the index, it is still
nice to fix for those times that I do.