search: add cookies to HTTP requests
Because no engine actually used it (well, one almost did, but I stopped
working on it because I couldn't break the bot defense) this code never
ended up being used when it was introduced in
0b852e4ee26d65451d68208bb38f486f83639749 and as such it never actually
worked properly.

I think in normal usage the cookies are to be added by
net/http.Client.Get/Post but we don't use them because we need to add a
bunch of headers onto the request to make our request seem more
legitimate.