partial uBlacklist ruleset support
This commit introduces partial support for blacklists following the
uBlacklist format.
It simultaneously deprecates the previous way of doing this in the
config file with `replace: ""`, as the blacklists provide a far nicer
way of doing this.

The current implementation relies heavily on regular expressions, which
is probably not ideal.
I'll fix this sooner or later.